refactor: centralize token iteration logic
