+++ src/integration/TokenStream.js
@@ -1,6 +1,7 @@
 import { Readable } from 'stream';
 import { CharStream } from '../lexer/CharStream.js';
 import { LexerEngine } from '../lexer/LexerEngine.js';
+import { tokenIterator } from './tokenUtils.js';
 
 /**
  * Create a Readable stream that emits tokens for syntax highlighting.
@@ -10,30 +11,16 @@ import { LexerEngine } from '../lexer/LexerEngine.js';
 export function createTokenStream(code, { errorRecovery = false } = {}) {
   const stream = new CharStream(code);
   const engine = new LexerEngine(stream, { errorRecovery });
-  let trivia = [];
-  let prev = null;
+  const iter = tokenIterator(engine);
   return new Readable({
     objectMode: true,
     read() {
-      // eslint-disable-next-line no-constant-condition
-      while (true) {
-        const tok = engine.nextToken();
-        if (!tok) {
-          if (prev) prev.trivia.trailing = trivia;
-          this.push(null);
-          return;
-        }
-        if (tok.type === 'WHITESPACE') {
-          trivia.push(tok);
-          continue;
-        }
-        tok.trivia.leading = trivia;
-        if (prev) prev.trivia.trailing = trivia;
-        trivia = [];
-        prev = tok;
-        this.push(tok);
-        break;
+      const { value, done } = iter.next();
+      if (done) {
+        this.push(null);
+        return;
       }
+      this.push(value);
     }
   });
 }
